SED-1: Redo the project structure to accomodate test
